backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.4.1",
    "django-ninja-jwt>=5.3.7",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
